reorganizing modules + decoupling alpss_main
