+++ src/alpss/analysis/spall.py
