+++ src/alpss/detection/spall_doi_finder.py
