+++ src/alpss/io/__init__.py
@@ -1,0 +1,1 @@
+from alpss.io.saving import save
