+++ src/alpss/plotting/__init__.py
@@ -1,0 +1,1 @@
+from alpss.plotting.plots import plot_results, plot_voltage
