+++ src/alpss/velocity/__init__.py
@@ -1,0 +1,3 @@
+from alpss.velocity.calculation import velocity_calculation
+from alpss.velocity.derivative import *
+from alpss.velocity.smoothing import *
